refactor: products api 절대경로 변경
